pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.coverage.run]
branch = true
relative_files = true
omit = [
    '.tox/*',
    'tests/*',
]


[tool.poetry]
name = "yptl"
version = "0.0.0"
description = "yaml pytorch lightning (yptl) is a python package that provides an executable to configure a model and define the training based on yaml-configuration files."
authors = ["Julian Bissantz <j.bissantz+github@gmail.com>"]
maintainers = ["Julian Bissantz <j.bissantz+github@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = 'https://github.com/chilluxer/yptl'


[tool.poetry.dependencies]
python = "^3.8"
pytorch-lightning = "^2.1.2"
torchvision = "^0.16.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.7.1"


[tool.poetry.group.linters]
optional = true


[tool.poetry.group.linters.dependencies]
ruff = "^0.1.5"


[tool.poetry.group.tests]
optional = true


[tool.poetry.group.tests.dependencies]
pytest = '^7.4.3'
pytest-cov = '^4.1.0'
tox = '^4.11.4'
tox-gh-actions = '^3.1.3'


[tool.poetry.urls]
"Tracker" = 'https://github.com/chilluxer/yptl/issues'
"Changelog" = 'https://github.com/chilluxer/yptl/blob/main/CHANGELOG.md'

[tool.poetry.scripts]
yptl = "yptl.scripts.yptl:main"

[tool.pytest.ini_options]
minversion = '6.0'
addopts = """
--doctest-modules \
--cov=./ \
--cov-append \
--cov-report html:tests/reports/coverage-html \
--cov-report xml:tests/reports/coverage.xml \
--ignore=docs/
"""
testpaths = [
    'tests',
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4

select = [
    'A',    # Builtins
    'ANN',  # Annotations
    'ARG',  # Unused arguments
    'B',    # Bugbear
    'BLE',  # Blind except
    'C4',   # Comprehensions
    'C90',  # mccabe
    'COM',  # Commas
    'D1',   # Undocumented public elements
    'D2',   # Docstring conventions
    'D3',   # Triple double quotes
    'D4',   # Docstring text format
    'DTZ',  # Datetimes
    'EM',   # Error messages
    'ERA',  # Commented-out code
    'EXE',  # Executable
    'F',    # Pyflakes
    'FA',   # __future__ annotations
    'FLY',  # F-strings
    # 'FURB', # Refurb
    'G',    # Logging format
    'I',    # Isort
    'ICN',  # Import conventions
    'INP',  # Disallow PEP-420 (Implicit namespace packages)
    'INT',  # gettext
    'ISC',  # Implicit str concat
    # 'LOG',  # Logging
    'N',    # PEP-8 Naming
    'NPY',  # Numpy
    'PERF', # Unnecessary performance costs
    'PGH',  # Pygrep hooks
    'PIE',  # Unnecessary code
    'PL',   # Pylint
    'PT',   # Pytest
    'PTH',  # Use Pathlib
    'PYI',  # Stub files
    'Q',    # Quotes
    'RET',  # Return
    'RUF',  # Ruff
    'RSE',  # Raise
    'S',    # Bandit
    'SIM',  # Code simplification
    'SLF',  # Private member access
    'SLOT', # __slots__
    'T10',  # Debugger
    'T20',  # Print
    'TCH',  # Type checking
    'TID',  # Tidy imports
    'TRY',  # Exception handling
    'UP',   # Pyupgrade
    'W',    # Warnings
    'YTT',  # sys.version
]
ignore = [
    'D203',    # One blank line before class docstring
    'D212',    # Multi-line summary first line
    'PLR0913', # Too many arguments
    'Q000',    # Single quotes found but double quotes preferred
    'ISC001',  # single-line-implicit-string-concatenation
    'COM812',  # missing-trailing-comma
    'ANN101',  # Type annotation for self in method 
]
ignore-init-module-imports = true
# preview = true
show-fixes = true
target-version = 'py38'


[tool.ruff.flake8-quotes]
docstring-quotes = 'double'
multiline-quotes = 'double'


[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.per-file-ignores]
# https://beta.ruff.rs/docs/rules/
'__init__.py' = ['F401','F403','F405',]
'tests/*' = ['ANN',
             'ARG',
             'INP001',
             'S101',
             'D',       # pydocstyle
]


[tool.ruff.pylint]
max-args = 15
max-branches = 20
max-returns = 10
max-statements = 80


[tool.ruff.flake8-tidy-imports]
ban-relative-imports = 'all'


[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py38, py39, py310, py311, pypy3
skip_missing_interpreters = true

[gh-actions]
python =
    3.8: py38
    3.9: py39
    3.10: py310
    3.11: py311

[testenv]
passenv = GITHUB_*
allowlist_externals = poetry

commands =
    poetry run pytest
    poetry run coverage report
"""
